+++ FE/client/src/components/header/top/TopNav.jsx
@@ -1,8 +1,8 @@
-import React, { useState, useEffect } from 'react';
+import React from 'react';
 import TopNavPopup from './TopNavPopup';
+import Login from './Login';
 import { makeStyles } from '@material-ui/core/styles';
 import { topNavData } from '../../../data/topNavData';
-import { getCookie, deleteCookie } from '../../../util/cookie';
 
 const useStyles = makeStyles({
     topNavWrap: {
@@ -50,32 +50,13 @@ const useStyles = makeStyles({
 function TopNav() {
     const { downloadApp, myPage, customerCenter } = topNavData;
     const classes = useStyles();
-    const [userData, setUserData] = useState(null);
-
-    const cookieKey = 'userId';
-    useEffect(() => {
-        const userId = getCookie(cookieKey);
-        setUserData(userId);
-    }, [userData]);
-
-    const logout = () => {
-        deleteCookie(cookieKey);
-        setUserData(null);
-    }
 
     return (
         <div className={classes.topNavWrap}>
             <div className={classes.topNavInner}>
                 <TopNavPopup title={downloadApp.title} contents={downloadApp.contents} />
                 <ul className={classes.topNavList}>
-                    {
-                        !userData ?
-                            <li className='singleList'><a href='https://github.com/login/oauth/authorize?client_id=8d92d01b11ba14d3d18f&scope=user%20public_repo'>로그인</a></li> :
-                            <>
-                                <span style={{ marginRight: '30px', cursor: 'auto' }}><span style={{ fontWeight: '600' }}>{userData}</span>님 반갑습니다!</span>
-                                <li className='singleList' onClick={logout}>로그아웃</li>
-                            </>
-                    }
+                    <Login />
                     <li className='singleList'>회원가입</li>
                     <li>
                         <TopNavPopup title={myPage.title} contents={myPage.contents} />
